addtocart

+++ app/src/main/java/com/example/foodapp/onChangeItem.java
@@ -1,5 +1,5 @@
 package com.example.foodapp;
 
 public interface onChangeItem {
-    void onPriceChange(int price);
+    void onPriceChange(double price);
 }
